refactor: 클래스 역할 및 구조별 패키지 리팩토링
