+++ src/main/java/racingcar/view/OutputView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
